Rename db folder to augur_db so it doesn't conflict with other naming'
