+++ augur_new/augur_db/models/base.py
