+++ augur_new/cli/config.py
@@ -9,11 +9,11 @@ import json
 import logging
 from pathlib import Path
 
-from augur.config import default_config, ENVVAR_PREFIX, CONFIG_HOME
-from augur.cli import initialize_logging
-from augur.logging import ROOT_AUGUR_DIRECTORY
+# from augur.config import default_config, ENVVAR_PREFIX, CONFIG_HOME
+# from augur.cli import initialize_logging
+# from augur.logging import ROOT_AUGUR_DIRECTORY
 
-from db.models import Config
+from augur_db.models import Config
 from tasks.task_session import TaskSession
 from augur_config import AugurConfig
