+++ augur_new/schema/alembic/env.py
@@ -4,8 +4,8 @@ from sqlalchemy import engine_from_config
 from sqlalchemy import pool
 
 from alembic import context
-from db.models.base import Base
-from db.engine import engine
+from augur_db.models.base import Base
+from augur_db.engine import engine
 
 # this is the Alembic Config object, which provides
 # access to the values within the .ini file in use.
